# appservice/src/main/resources/static/swagger/swagger-spec.yaml
swagger: '2.0'
info:
  title: Yona app service API
  description: The API of the service backing the Yona app
  version: 1.0.0
produces:
  - application/json
parameters:
  userID:
    name: userID
    in: path
    description: The ID of the user
    required: true
    type: string
  yonaPassword:
    name: Yona-Password
    in: header
    description: The device password
    required: true
    type: string
  yonaPasswordOptional:
    name: Yona-Password
    in: header
    description: The device password
    required: false
    type: string
...
  requestingUserId:
    name: requestingUserId
    in: query
    description: The user ID of the user GETting this user entity. Either equal to the ID of the user being fetched or the user ID of one of their buddies.
    required: true
    type: string
  requestingDeviceId:
    name: requestingDeviceId
    in: query
    description: The ID of the device from which the request originates. Several operations take this query parameter to enable returning a device-specific URL for the open-app event and to post the app activities 
    required: false
    type: string
  tempPassword:
    name: tempPassword
    in: query
    description: The temporary password that was included in the invitation e-mail when a user invited a non-Yona user as buddy
    required: false
    type: string
...
paths:
  /users/:
    post:
      summary: Add a new user
      description: Registers a new user
      parameters:
        - name: overwriteUserConfirmationCode
          in: query
          description: Pass the confirmation code here when trying to overwrite the account in case of a device loss
          required: false
          type: string
        - in: body
          name: body
          description: New user
          required: true
          schema:
            $ref: '#/definitions/PostPutUser'
      tags:
        - User
      responses:
        '201':
          description: The created user
          schema:
            $ref: '#/definitions/UserWithPrivateData'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/ConfirmationError'
  '/users/{userID}':
    get:
      summary: Get a user
      description: Fetches the user identified by the given ID
      parameters:
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/yonaPasswordOptional"
        - $ref: "#/parameters/tempPassword"
        - $ref: "#/parameters/requestingUserId"
        - $ref: "#/parameters/requestingDeviceId"
      tags:
        - User
      responses:
        '200':
          description: The requested user. The set of returned properties depends on the query parameter requestingUserId.
          schema:
            $ref: '#/definitions/UserWithPrivateData'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Update a user
      description: Fetches the user identified by the given ID
      parameters:
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/requestingDeviceId"
        - $ref: "#/parameters/yonaPasswordOptional"
        - $ref: "#/parameters/tempPassword"
        - in: body
          name: body
          description: New user
          required: true
          schema:
            $ref: '#/definitions/PostPutUser'
      tags:
        - User
      responses:
        '200':
          description: The updated user
          schema:
            $ref: '#/definitions/UserWithPrivateData'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete a user
      description: Deletes the user with the given ID
      parameters:
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/yonaPassword"
        - name: message
          in: query
          description: A message of the user to their buddies, explaining why they unsubscribe from Yona
          required: false
          type: string
      tags:
        - User
      responses:
        '204':
          description: The user is deleted.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/users/{userID}/photo':
    put:
      summary: Upload a user photo
      description: Adds or replaces the user photo for this user. A multipart/form-data request is expected with a "file" field containing an image. The image is expected to be resized already.
      consumes: 
        - multipart/form-data
      parameters:
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/yonaPassword"
        - in: formData
          name: file
          type: file
          description: The image to upload. It should be resized already.
      tags:
        - User photo
      responses:
        '200':
          description: The user photo added
          schema:
            $ref: '#/definitions/UserPhoto'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Remove a user photo
      description: Removes the photo association of this user. It is still downloadable for some time.
      parameters:
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/yonaPassword"
      tags:
        - User photo
      responses:
        '204':
          description: The user photo is deleted.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'               
  '/userPhotos/{userPhotoID}':
    get:
      summary: Get the user photo
      description: Fetches the user photo by ID.
      parameters:
        - name: userPhotoID
          in: path
          description: The ID of the user photo
          required: true
          type: string
      tags:
        - User photo
      responses:
        '200':
          description: The user photo in image/png format
          schema:
            type: file
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/users/{userID}/pinResetRequest/request':
    post:
      summary: Request pin reset confirmation code
      description: Post an empty body to this URL to trigger the generation of a pin reset confirmation code
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
      tags:
        - Pin reset
      responses:
        '200':
          description: Request registered successfully
          schema:
            type: object
            properties:
              delay:
                type: string
                description: Delay till the confirmation code will be sent by text message. The format complies with the ISO-8601 duration format (PnDTnHnMn.nS).
                example: P1D
            required:
              - delay
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/users/{userID}/pinResetRequest/verify':
    post:
      summary: Verify pin reset confirmation code
      description: Post the pin reset confirmation code here for verification
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
        - in: body
          name: body
          description: Message with confirmation code
          required: true
          schema:
            $ref: '#/definitions/ConfirmationCode'
      tags:
        - Pin reset
      responses:
        '204':
          description: Confirmation code correct
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ConfirmationError'
  '/users/{userID}/pinResetRequest/resend':
    post:
      summary: Request resend of a pin reset confirmation code
      description: Post an empty request to get a new pin reset confirmation code texted to the user
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
      tags:
        - Pin reset
      responses:
        '204':
          description: The request is successfully handled
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/users/{userID}/pinResetRequest/clear':
    post:
      summary: Clear pin reset confirmation code
      description: Post an empty body to this URL to clear the pin reset confirmation code
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
      tags:
        - Pin reset
      responses:
        '204':
          description: Request cleared successfully
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Curies:
    type: object
    properties:
      curies:
        type: array
        description: A set of curies (Compact URIs). For Yona, there is just one curie.
        items:
          type: object
          properties:
            name:
              type: string
              description: The curie name. The prefix of a curied link will be used to find the curie by name
              example: yona
            href:
              type: string
              description: The URL corresponding to this curie.
              example: http://dev.yona.nu/rels/{rel}
            templated:
              type: string
              format: boolean
              description: Whether this curie is templated (always true for Yona).
              example: true
          required:
            - name
            - href
            - templated
  SelfLink:
    type: object
    properties:
      self:
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Identifies this same resource
    required:
      - self
  EditLink:
    type: object
    properties:
      edit:
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Through this link, this resource can be edited or deleted. Note that this link is not always present. Certain resources can never be edited. Some others can only be edited or deleted in a specific state. Unprocessed messages for example cannot be deleted.
...
  WeeklyActivityReportsLink:
    type: object
    properties:
      "yona:weeklyActivityReports":
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Refers to the weekly activity reports
  DailyActivityReportsLink:
    type: object
    properties:
      "yona:dailyActivityReports":
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Refers to the daily activity reports
...
  DeviceOperatingSystem:
    type: string
    description: The operating system of the device running the Yona app
    enum:
    - ANDROID
    - IOS
    - UNKNOWN
    example: ANDROID
  DeviceAppVersion:
    type: string
    description: The version of the Yona app running on this device. This is the marketing version, a free text string.
    example: 1.1 build 83
  DeviceAppVersionCode:
    type: integer
    description: The version code of the Yona app running on this device. This is a positive, monotonic increasing number, across releases
    example: 31
  FirebaseInstanceId:
    type: string
    description: The Instance ID provided by Firebase to the app
    example: d3cIznsu5VQ:APA91bGWLq7xBK1RDkpGURdliHb-S_nCBLqYnXhEWfGnItP_qGDZ6f2EF1mB66yHdBiicggV7APIWwkQXTUq_zJgwPkJtvcdqpUphYN7p8E8Sq02_ErljVApX8-n9-nvVxiyqmUg9ALZ
...
  BaseUser:
    type: object
    properties:
      firstName:
        type: string
        description: First name.
        example: Richard
      lastName:
        type: string
        description: Last name.
        example: Quin
      mobileNumber:
        type: string
        description: 'Mobile number, unique among all users.'
        example: "+31612345678"
      nickname:
        type: string
        description: Nickname known among buddies.
        example: RQ
    required:
    - firstName
    - lastName
    - mobileNumber
    - nickname
  PrivateUserData:
    type: object
    properties:
      yonaPassword:
        type: string
        description: The Yona password for subsequent requests
        example: AES:128:YXNkZnNhZGZnMnrp1ORubw==
  PostPutUser:
    allOf:
      - $ref: '#/definitions/BaseUser'
      - type: object
        properties:
          deviceName:
            type: string
            description: The name of the device on which the user signs up for Yona.  The name must be at most 20 characters and should not contain a colon. Not supported for a regular user update.
            example: My S8
          deviceOperatingSystem:
            $ref: '#/definitions/DeviceOperatingSystem'
          deviceAppVersion:
            $ref: '#/definitions/DeviceAppVersion'
          deviceAppVersionCode:
            $ref: '#/definitions/DeviceAppVersionCode'
          deviceFirebaseInstanceId:
            $ref: '#/definitions/FirebaseInstanceId'
  UserWithPrivateData:
    allOf:
      - $ref: '#/definitions/BaseUser'
      - $ref: '#/definitions/PrivateUserData'
      - type: object
        properties:
          creationTime:
            type: string
            format: dateTime
            description: The creation time of this user. Not available for buddies.
            example: 2016-04-26T20:08:55.365+0200
          appLastOpenedDate:
            type: string
            format: date
            description: The date the user last opened the Yona app
            example: 2016-04-26
          lastMonitoredActivityDate:
            type: string
            format: date
            description: The date of the last monitored activity.  Not available for buddies.
            example: 2016-04-26
          _embedded:
            type: object
            properties:
              "yona:goals":
                type: object
                description: The goals of this user.
                properties:
                  _embedded:
                    type: object
                    properties:
                      "yona:goals":
                        type: array
                        items:
                          $ref: "#/definitions/Goal"
                  _links:
                    allOf:
                      - $ref: '#/definitions/SelfLink'
                      - $ref: '#/definitions/EditLink'
              "yona:buddies":
                type: object
                description: The buddies of this user.
                properties:
                  _embedded:
                    type: object
                    properties:
                      "yona:buddies":
                        type: array
                        items:
                          $ref: "#/definitions/Buddy"
                  _links:
                    allOf:
                      - $ref: '#/definitions/SelfLink'
                      - $ref: '#/definitions/EditLink'
              "yona:devices":
                type: object
                description: The devices of this user.
                properties:
                  _embedded:
                    type: object
                    properties:
                      "yona:devices":
                        type: array
                        items:
                          $ref: "#/definitions/UserDevice"
                  _links:
                    allOf:
                      - $ref: '#/definitions/SelfLink'
                      - $ref: '#/definitions/EditLink'
          _links:
            allOf:
              - $ref: '#/definitions/Curies'
              - $ref: '#/definitions/SelfLink'
              - $ref: '#/definitions/EditLink'
              - $ref: '#/definitions/WeeklyActivityReportsLink'
              - $ref: '#/definitions/DailyActivityReportsLink'
              - $ref: '#/definitions/DailyActivityReportsWithBuddiesLink'
              - $ref: '#/definitions/MessagesLink'
              - $ref: '#/definitions/UserPhotoLink'
              - $ref: '#/definitions/EditUserPhotoLink'
              - type: object
                properties:
                  "yona:confirmMobileNumber":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: If this link is present, the mobile number of the user must be confirmed by posting the confirmation to this URL.
                  "yona:resendMobileNumberConfirmationCode":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: This link allows requesting a resend of the mobile number confirmation code.
                  "yona:requestPinReset":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: This link allows requesting a pin reset confirmation code.
                  "yona:verifyPinReset":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: This link allows verifying the pin reset confirmation code.
                  "yona:resendPinResetConfirmationCode":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: This link allows requesting a resend of a new pin reset confirmation code.
                  "yona:clearPinReset":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: This link allows clearing the pin reset confirmation code.
                  "yona:newDeviceRequest":
                    type: object
                    required:
                      - href
                    properties:
                      href:
                        type: string
                        description: Create, get or delete a "new device request" through this link
        required:
          - _embedded
          - _links
...
  ConfirmationCode:
    type: object
    properties:
      code:
        type: string
        example: 1234
    required:
      - code
  UserPhoto:
    type: object
    properties:
      _links:
        allOf:
          - $ref: '#/definitions/Curies'
          - $ref: '#/definitions/SelfLink'
    required:
      - _links
...
  Goal:
    type: object
    discriminator: "@type"
    properties:
      "@type":
        type: string
        enum:
        - BudgetGoal
        - TimeZoneGoal
        description: The type of goal
      _links:
        allOf:
          - $ref: '#/definitions/Curies'
          - $ref: '#/definitions/SelfLink'
          - $ref: '#/definitions/EditLink'
          - $ref: '#/definitions/ActivityCategoryLink'
      creationTime:
        type: string
        format: dateTime
        description: The creation time of this goal
        example: 2016-04-26T20:08:55.365+0200
      historyItem:
        type: boolean
        example: true
        description: Indicates whether the goal is retained for historical reasons only. Such goals should not feature in the user's goal list but should be used when displaying activity information.
    required:
      - "@type"
      - _links
      - creationTime
...
  Buddy:
    allOf:
      - $ref: '#/definitions/BaseBuddy'
      - type: object
        properties:
          lastStatusChangeTime:
            type: string
            format: dateTime
            description: The time of the last status change (request/accept/reject) of this buddy
            example: 2016-02-23T21:28:58.556+0000
          lastMonitoredActivityDate:
            type: string
            format: date
            description: The date of the last monitored activity of this buddy
            example: 2016-04-26
          _embedded:
            type: object
            properties:
              "yona:user":
                  $ref: "#/definitions/UserWithOnlyPublicData"
            required:
              - "yona:user"
          _links:
            allOf:
              - $ref: '#/definitions/Curies'
              - $ref: '#/definitions/SelfLink'
              - $ref: '#/definitions/EditLink'
              - $ref: '#/definitions/WeeklyActivityReportsLink'
              - $ref: '#/definitions/DailyActivityReportsLink'
        required:
          - lastStatusChangeTime
          - _embedded
          - _links
...
  Error:
    type: object
    properties:
      code:
        type: string
        description: The error code
        example: error.user.mobile.number.invalid
      message:
        type: string
        description: A localized error message
        example: The mobile number ''+31 6 12345678'' is invalid. It must start with a + sign, with no spaces between the digits
      correlationId:
        type: string
        description: ID that correlates the server log messages to this error
        example: 463ac35c9f6413ad48485a3953bb6124
  ConfirmationError:
    allOf:
      - $ref: '#/definitions/Error'
      - type: object
        properties:
          remainingAttempts:
            type: integer
            description: The number of remaining attempts. Only available upon a failed confirmation attempt
            example: 4